minor refactoring of assembler module
